Add enter from verbotten keys
I've added the enter key as a verbotten key. Since there's two places
that track these keys, I've added it to the config to make it easier to
update.

I also found some code that wasn't doing much except messing with the
focus of cursor when editing numbers in the explore rates tool.