Adding new color palette
I've updated the color palette and updated all the colors in the project
accordingly. One legacy color still remains because it is used by the
version of Capital Framework we are using and can be safely removed when
we update the project. It has been clearly labeled as a legacy variable.