Added a brand header for Owning a Home pages (#693)
* Added a brand header for Owning a Home pages

Created an include for a brand header to be included on all Owning a
Home pages. Right now this will be only on non-wagtail pages, but any
advice that can make this portable to wagtail is appreciated.

* Fixing the brand header to use abs positioning

Since the solution with the grid was a little too flexible and left too
much space at some breakpoints. Trying out a solution with absolute
positioning as an alternative.